Move HSTX pins to a submodule
